button testing

+++ src/addresableLED.cpp
@@ -1,7 +1,7 @@
 #include "addresableLED.h"
 #include <FastLED.h>
 
-void AddresableLED::Duha(CRGB addLEDs[],int numberOfLEDs, int HUEIncrementor)
+void AddresableLED::Duha(CRGB addLEDs[],uint8_t numberOfLEDs, uint8_t HUEIncrementor)
 {
     fill_rainbow(addLEDs, numberOfLEDs, 0, HUEIncrementor);
 }
