+++ src/addresableLED.h
@@ -2,6 +2,6 @@
 class AddresableLED
 {
     public:
-    void Duha(CRGB addLEDs[],int numberOfLEDs, int HUEIncrementor);
+    void Duha(CRGB addLEDs[],uint8_t numberOfLEDs, uint8_t HUEIncrementor);
     
 };
